backend-main/requirements.txt
alembic==1.10.2
aniso8601==9.0.1
attrs==22.2.0
chardet==3.0.4
click==8.1.3
colorama==0.4.6
entrypoints==0.3
flake8==3.7.9
Flask==2.1.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
gunicorn==20.0.4
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.6.1
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==7.2.2
pytest-flask==1.0.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.23.0
shortuuid==1.0.1
six==1.16.0
SQLAlchemy==1.4.0
SQLAlchemy-Utils==0.40.0
typing_extensions==4.5.0
urllib3==1.25.11
Werkzeug==2.2.3
wincertstore==0.2
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
Flask-Cors==3.0.10
